Extract In-Browser Editor to EditorBundle
