+++ src/Sculpin/Bundle/EditorBundle/DependencyInjection/Configuration.php
@@ -1,0 +1,35 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is a part of Sculpin.
+ *
+ * (c) Dragonfly Development Inc.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sculpin\Bundle\EditorBundle\DependencyInjection;
+
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+
+/**
+ * @author Kevin Boyd <kevin@whateverthing.com>
+ */
+class Configuration implements ConfigurationInterface
+{
+    /**
+    * {@inheritdoc}
+    */
+    public function getConfigTreeBuilder(): TreeBuilder
+    {
+        $treeBuilder = new TreeBuilder('sculpin_editor');
+
+        $rootNode = $treeBuilder->getRootNode();
+
+        return $treeBuilder;
+    }
+}
