+++ src/Sculpin/Bundle/EditorBundle/Resources/css/editor.css
