+++ src/Sculpin/Bundle/EditorBundle/Resources/js/editor.js
